+++ app/(auth)/(register)/terms/marketing/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import Link from "next/link";
-import { Button } from "@/components/ui/Button";
+import { Button } from "@/components/ui/button";
 import { FileText, HelpCircle } from "lucide-react";
 
 /**
