+++ app/(auth)/(register)/terms/privacy/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import Link from "next/link";
-import { Button } from "@/components/ui/Button";
+import { Button } from "@/components/ui/button";
 import { AlertCircle, FileText, HelpCircle, Shield } from "lucide-react";
 
 /**
