+++ components/auth/register/step2-business.tsx
@@ -5,11 +5,11 @@ import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 
 import StepContainer from "./step-container";
-import { Button } from "@/components/ui/Button";
-import { Input } from "@/components/ui/Input";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
 import { RegisterFormData, Step2Data } from "@/types/register";
-import { useBusinessLicenseUpload } from "@/hooks/api/register/useBusinessLicenseUpload";
-import { useIdDuplicateCheck } from "@/hooks/api/register/useIdDuplicateCheck";
+import { useBusinessLicenseUpload } from "@/hooks/api/register/use-business-license-upload";
+import { useIdDuplicateCheck } from "@/hooks/api/register/use-id-duplicate-check";
 
 interface Step2Props {
   onNext: (data: Step2Data) => void;
