+++ components/dashboard/help-section.tsx
@@ -1,4 +1,4 @@
-import { Button } from "@/components/ui/Button";
+import { Button } from "@/components/ui/button";
 
 export const HelpSection = () => {
   return (
