+++ components/dashboard/menu/modal/menu-option-editor.tsx
@@ -1,6 +1,6 @@
 import { Plus, Trash } from "lucide-react";
-import { Button } from "@/components/ui/Button";
-import { Input } from "@/components/ui/Input";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
 import { MenuOptionGroup } from "@/app/api/dashboard/menu/route";
 
 /**
