+++ components/dashboard/settings/merchant-info.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Input } from "@/components/ui/Input";
+import { Input } from "@/components/ui/input";
 import { MerchantInfoResponse } from "@/app/api/merchant/route";
 
 interface MerchantInfoProps {
