+++ components/ui/modal.tsx
@@ -1,7 +1,7 @@
 import React, { forwardRef, useEffect, useState } from "react";
 import { cn } from "@/lib/utils";
 import { X } from "lucide-react";
-import { Button } from "./Button";
+import { Button } from "./button";
 
 export interface ModalProps
   extends Omit<React.HTMLAttributes<HTMLDivElement>, "title"> {
