+++ hooks/api/dashboard/menu/use-delete-menu.ts
