+++ hooks/api/dashboard/menu/use-update-menu-availability.ts
