+++ hooks/api/register/use-id-duplicate-check.ts
