+++ hooks/api/register/use-merchant-image-upload.ts
