+++ stories/badge.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
-import { Badge } from "@/components/ui/Badge";
+import { Badge } from "@/components/ui/badge";
 import { CheckCheck, Clock, Crown, Shield, Star } from "lucide-react";
 
 const meta = {
