+++ stories/input.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import type { Meta, StoryObj } from "@storybook/react";
-import { Input } from "@/components/ui/Input";
+import { Input } from "@/components/ui/input";
 import {
   AlertCircle,
   Calendar,
